// src/movies/types.ts
declare type MovieData = {
  "#TITLE": string;
  "#YEAR": number;
  "#IMDB_ID": string;
  "#RANK": number;
  "#ACTORS": string;
  "#AKA": string;
  "#IMDB_URL": string;
  "#IMDB_IV": string;
  "#IMG_POSTER": string;
  photo_width: number;
  photo_height: number;
};

export declare type ResponseMovieData = {
  ok: boolean;
  description: MovieData[];
  error_code: string;
};
export declare type Movie = {
  title: string;
  year: number;
  imdbId: string;
  rank: number;
  actors: string;
  aka: string;
  imdbUrl: string;
  imdbIv: string;
  imgPoster: string;
  photo_width: number;
  photo_height: number;
};
export declare type SearchResponse = {
  movies: Movie[];
  ok: boolean;
  error_code: string;
};

/**
 * Details Data Type
 */
export declare type DetailsData = {
  short: Short;
  imdbId: string;
  top: Top;
  main: Main;
  fake: Fake;
  storyLine: StoryLine;
};
export declare type ResponseDetailsData = {
  ok: boolean;
  items: DetailsData[];
  error_code: string;
};
export declare type Details = {
  url: string;
  name: string;
  description: string;
  imageUrl: string;
  review: Review;
  genres: string[];
  keywords: string;
  actors: Actor[];
  director: Director[];
  duration: string;
  id: string;
  rating: string;
};
export declare type DetailsResponse = {
  details: Details;
  ok: boolean;
  error_code: string;
};

export declare type RandomMoviesResponse = {
  movies: Movie[];
  ok: boolean;
  error_code: string;
};

export interface Short {
  "@context": string;
  "@type": string;
  url: string;
  name: string;
  alternateName: string;
  image: string;
  description: string;
  review: Review;
  aggregateRating: AggregateRating;
  contentRating: string;
  genre: string[];
  datePublished: string;
  keywords: string;
  trailer: Trailer;
  actor: Actor[];
  director: Director[];
  creator: Creator[];
  duration: string;
}

export interface Review {
  "@type": string;
  itemReviewed: ItemReviewed;
  author: Author;
  dateCreated: string;
  inLanguage: string;
  name: string;
  reviewBody: string;
  reviewRating: ReviewRating;
}

export interface ItemReviewed {
  "@type": string;
  url: string;
}

export interface Author {
  "@type": string;
  name: string;
}

export interface ReviewRating {
  "@type": string;
  worstRating: number;
  bestRating: number;
  ratingValue: number;
}

export interface AggregateRating {
  "@type": string;
  ratingCount: number;
  bestRating: number;
  worstRating: number;
  ratingValue: number;
}

export interface Trailer {
  "@type": string;
  name: string;
  embedUrl: string;
  thumbnail: Thumbnail;
  thumbnailUrl: string;
  url: string;
  description: string;
  duration: string;
  uploadDate: string;
}

export interface Thumbnail {
  "@type": string;
  contentUrl: string;
}

export interface Actor {
  "@type": string;
  url: string;
  name: string;
}

export interface Director {
  "@type": string;
  url: string;
  name: string;
}

export interface Creator {
  "@type": string;
  url: string;
  name?: string;
}

export interface Top {
  id: string;
  productionStatus: ProductionStatus;
  canHaveEpisodes: boolean;
  series: any;
  titleText: TitleText;
  titleType: TitleType;
  originalTitleText: OriginalTitleText;
  certificate: Certificate;
  releaseYear: ReleaseYear;
  releaseDate: ReleaseDate;
  runtime: Runtime;
  canRate: CanRate;
  ratingsSummary: RatingsSummary;
  meterRanking: MeterRanking;
  primaryImage: PrimaryImage;
  images: Images;
  videos: Videos;
  primaryVideos: PrimaryVideos;
  externalLinks: ExternalLinks;
  metacritic: Metacritic;
  keywords: Keywords;
  genres: Genres;
  plot: Plot;
  plotContributionLink: PlotContributionLink;
  credits: Credits;
  principalCredits: PrincipalCredit[];
  reviews: Reviews;
  criticReviewsTotal: CriticReviewsTotal;
  triviaTotal: TriviaTotal;
  engagementStatistics: EngagementStatistics;
  subNavCredits: SubNavCredits;
  subNavReviews: SubNavReviews;
  subNavTrivia: SubNavTrivia;
  subNavFaqs: SubNavFaqs;
  subNavTopQuestions: SubNavTopQuestions;
  titleGenres: TitleGenres;
  meta: Meta;
  castPageTitle: CastPageTitle;
  creatorsPageTitle: any[];
  directorsPageTitle: DirectorsPageTitle[];
  countriesOfOrigin: CountriesOfOrigin;
  production: Production;
  featuredReviews: FeaturedReviews;
  __typename: string;
}

export interface ProductionStatus {
  currentProductionStage: CurrentProductionStage;
  productionStatusHistory: ProductionStatusHistory[];
  restriction: any;
  __typename: string;
}

export interface CurrentProductionStage {
  id: string;
  text: string;
  __typename: string;
}

export interface ProductionStatusHistory {
  status: Status;
  __typename: string;
}

export interface Status {
  id: string;
  text: string;
  __typename: string;
}

export interface TitleText {
  text: string;
  __typename: string;
}

export interface TitleType {
  displayableProperty: DisplayableProperty;
  text: string;
  id: string;
  isSeries: boolean;
  isEpisode: boolean;
  categories: Category[];
  canHaveEpisodes: boolean;
  __typename: string;
}

export interface DisplayableProperty {
  value: Value;
  __typename: string;
}

export interface Value {
  plainText: string;
  __typename: string;
}

export interface Category {
  value: string;
  __typename: string;
}

export interface OriginalTitleText {
  text: string;
  __typename: string;
}

export interface Certificate {
  rating: string;
  __typename: string;
}

export interface ReleaseYear {
  year: number;
  endYear: any;
  __typename: string;
}

export interface ReleaseDate {
  day: number;
  month: number;
  year: number;
  __typename: string;
}

export interface Runtime {
  seconds: number;
  displayableProperty: DisplayableProperty2;
  __typename: string;
}

export interface DisplayableProperty2 {
  value: Value2;
  __typename: string;
}

export interface Value2 {
  plainText: string;
  __typename: string;
}

export interface CanRate {
  isRatable: boolean;
  __typename: string;
}

export interface RatingsSummary {
  aggregateRating: number;
  voteCount: number;
  __typename: string;
}

export interface MeterRanking {
  currentRank: number;
  rankChange: RankChange;
  __typename: string;
}

export interface RankChange {
  changeDirection: string;
  difference: number;
  __typename: string;
}

export interface PrimaryImage {
  id: string;
  width: number;
  height: number;
  url: string;
  caption: Caption;
  __typename: string;
}

export interface Caption {
  plainText: string;
  __typename: string;
}

export interface Images {
  total: number;
  edges: Edge[];
  __typename: string;
}

export interface Edge {
  node: Node;
  __typename: string;
}

export interface Node {
  id: string;
  __typename: string;
}

export interface Videos {
  total: number;
  __typename: string;
}

export interface PrimaryVideos {
  edges: Edge2[];
  __typename: string;
}
// ...
export interface ExternalLinks {
  total: number;
  __typename: string;
}

export interface Metacritic {
  metascore: Metascore;
  __typename: string;
}

export interface Metascore {
  score: number;
  __typename: string;
}

export interface Keywords {
  total: number;
  edges: Edge3[];
  __typename: string;
}

export interface Edge3 {
  node: Node3;
  __typename: string;
}

export interface Node3 {
  text: string;
  __typename: string;
}

export interface Genres {
  genres: Genre[];
  __typename: string;
}

export interface Genre {
  text: string;
  id: string;
  __typename: string;
}

export interface Plot {
  plotText: PlotText;
  language: Language;
  __typename: string;
}

export interface PlotText {
  plainText: string;
  __typename: string;
}

export interface Language {
  id: string;
  __typename: string;
}

export interface PlotContributionLink {
  url: string;
  __typename: string;
}

export interface Credits {
  total: number;
  __typename: string;
}

export interface PrincipalCredit {
  totalCredits: number;
  category: Category2;
  credits: Credit[];
  __typename: string;
}

export interface Category2 {
  text: string;
  id: string;
  __typename: string;
}

export interface Credit {
  name: Name2;
  attributes: any;
  __typename: string;
}

export interface Name2 {
  nameText: NameText;
  id: string;
  __typename: string;
}

export interface NameText {
  text: string;
  __typename: string;
}

export interface Reviews {
  total: number;
  __typename: string;
}

export interface CriticReviewsTotal {
  total: number;
  __typename: string;
}

export interface TriviaTotal {
  total: number;
  __typename: string;
}

export interface EngagementStatistics {
  watchlistStatistics: WatchlistStatistics;
  __typename: string;
}

export interface WatchlistStatistics {
  displayableCount: DisplayableCount;
  __typename: string;
}

export interface DisplayableCount {
  text: string;
  __typename: string;
}

export interface SubNavCredits {
  total: number;
  __typename: string;
}

export interface SubNavReviews {
  total: number;
  __typename: string;
}

export interface SubNavTrivia {
  total: number;
  __typename: string;
}

export interface SubNavFaqs {
  total: number;
  __typename: string;
}

export interface SubNavTopQuestions {
  total: number;
  __typename: string;
}

export interface TitleGenres {
  genres: Genre2[];
  __typename: string;
}

export interface Genre2 {
  genre: Genre3;
  __typename: string;
}

export interface Genre3 {
  text: string;
  __typename: string;
}

export interface Meta {
  canonicalId: string;
  publicationStatus: string;
  __typename: string;
}

export interface CastPageTitle {
  edges: Edge4[];
  __typename: string;
}

export interface Edge4 {
  node: Node4;
  __typename: string;
}

export interface Node4 {
  name: Name3;
  __typename: string;
}

export interface Name3 {
  id: string;
  nameText: NameText2;
  __typename: string;
}

export interface NameText2 {
  text: string;
  __typename: string;
}

export interface DirectorsPageTitle {
  credits: Credit2[];
  __typename: string;
}

export interface Credit2 {
  name: Name4;
  __typename: string;
}

export interface Name4 {
  id: string;
  nameText: NameText3;
  __typename: string;
}

export interface NameText3 {
  text: string;
  __typename: string;
}

export interface CountriesOfOrigin {
  countries: Country[];
  __typename: string;
}

export interface Country {
  id: string;
  __typename: string;
}

export interface Production {
  edges: Edge5[];
  __typename: string;
}

export interface Edge5 {
  node: Node5;
  __typename: string;
}

export interface Node5 {
  company: Company;
  __typename: string;
}

export interface Company {
  id: string;
  companyText: CompanyText;
  __typename: string;
}

export interface CompanyText {
  text: string;
  __typename: string;
}

export interface FeaturedReviews {
  edges: Edge6[];
  __typename: string;
}
// ...
export interface Main {
  id: string;
  wins: Wins;
  nominations: Nominations;
  prestigiousAwardSummary: PrestigiousAwardSummary;
  ratingsSummary: RatingsSummary2;
  episodes: any;
  videos: Videos2;
  videoStrip: VideoStrip;
  titleMainImages: TitleMainImages;
  productionStatus: ProductionStatus2;
  primaryImage: PrimaryImage2;
  imageUploadLink: ImageUploadLink;
  titleType: TitleType2;
  cast: Cast;
  creators: any[];
  directors: Director2[];
  writers: Writer[];
  isAdult: boolean;
  moreLikeThisTitles: MoreLikeThisTitles;
  triviaTotal: TriviaTotal2;
  trivia: Trivia;
  goofsTotal: GoofsTotal;
  goofs: Goofs;
  quotesTotal: QuotesTotal;
  quotes: Quotes;
  crazyCredits: CrazyCredits;
  alternateVersions: AlternateVersions;
  connections: Connections;
  soundtrack: Soundtrack;
  titleText: TitleText6;
  originalTitleText: OriginalTitleText6;
  releaseYear: ReleaseYear5;
  reviews: Reviews2;
  featuredReviews: FeaturedReviews2;
  canRate: CanRate3;
  iframeAddReviewLink: IframeAddReviewLink;
  topQuestions: TopQuestions;
  faqs: Faqs;
  releaseDate: ReleaseDate2;
  countriesOfOrigin: CountriesOfOrigin2;
  detailsExternalLinks: DetailsExternalLinks;
  spokenLanguages: SpokenLanguages;
  akas: Akas;
  filmingLocations: FilmingLocations;
  production: Production2;
  companies: Companies;
  productionBudget: ProductionBudget;
  lifetimeGross: LifetimeGross;
  openingWeekendGross: OpeningWeekendGross;
  worldwideGross: WorldwideGross;
  technicalSpecifications: TechnicalSpecifications;
  runtime: Runtime5;
  series: any;
  canHaveEpisodes: boolean;
  contributionQuestions: ContributionQuestions;
  __typename: string;
}

export interface Wins {
  total: number;
  __typename: string;
}

export interface Nominations {
  total: number;
  __typename: string;
}

export interface PrestigiousAwardSummary {
  nominations: number;
  wins: number;
  award: Award;
  __typename: string;
}

export interface Award {
  text: string;
  id: string;
  event: Event;
  __typename: string;
}

export interface Event {
  id: string;
  __typename: string;
}

export interface RatingsSummary2 {
  topRanking: TopRanking;
  __typename: string;
}

export interface TopRanking {
  id: string;
  text: Text2;
  rank: number;
  __typename: string;
}

export interface Text2 {
  value: string;
  __typename: string;
}

export interface Videos2 {
  total: number;
  __typename: string;
}

export interface VideoStrip {
  edges: Edge7[];
  __typename: string;
}
// ...
export interface TitleMainImages {
  total: number;
  edges: Edge8[];
  __typename: string;
}

export interface Edge8 {
  node: Node8;
  __typename: string;
}

export interface Node8 {
  id: string;
  url: string;
  caption: Caption2;
  height: number;
  width: number;
  __typename: string;
}

export interface Caption2 {
  plainText: string;
  __typename: string;
}

export interface ProductionStatus2 {
  currentProductionStage: CurrentProductionStage2;
  productionStatusHistory: ProductionStatusHistory2[];
  restriction: any;
  __typename: string;
}

export interface CurrentProductionStage2 {
  id: string;
  text: string;
  __typename: string;
}

export interface ProductionStatusHistory2 {
  status: Status2;
  __typename: string;
}

export interface Status2 {
  id: string;
  text: string;
  __typename: string;
}

export interface PrimaryImage2 {
  id: string;
  __typename: string;
}

export interface ImageUploadLink {
  url: string;
  __typename: string;
}

export interface TitleType2 {
  id: string;
  canHaveEpisodes: boolean;
  __typename: string;
}

export interface Cast {
  edges: Edge9[];
  __typename: string;
}
// ...
export interface Director2 {
  totalCredits: number;
  category: Category4;
  credits: Credit3[];
  __typename: string;
}

export interface Category4 {
  text: string;
  __typename: string;
}

export interface Credit3 {
  name: Name7;
  attributes: any;
  __typename: string;
}

export interface Name7 {
  id: string;
  nameText: NameText5;
  __typename: string;
}

export interface NameText5 {
  text: string;
  __typename: string;
}

export interface Writer {
  totalCredits: number;
  category: Category5;
  credits: Credit4[];
  __typename: string;
}

export interface Category5 {
  text: string;
  __typename: string;
}

export interface Credit4 {
  name: Name8;
  attributes: any;
  __typename: string;
}

export interface Name8 {
  id: string;
  nameText: NameText6;
  __typename: string;
}

export interface NameText6 {
  text: string;
  __typename: string;
}

export interface MoreLikeThisTitles {
  edges: Edge10[];
  __typename: string;
}
// ...
export interface TriviaTotal2 {
  total: number;
  __typename: string;
}

export interface Trivia {
  edges: Edge11[];
  __typename: string;
}

export interface Edge11 {
  node: Node11;
  __typename: string;
}

export interface Node11 {
  text: Text3;
  trademark: any;
  relatedNames: any;
  __typename: string;
}

export interface Text3 {
  plaidHtml: string;
  __typename: string;
}

export interface GoofsTotal {
  total: number;
  __typename: string;
}

export interface Goofs {
  edges: Edge12[];
  __typename: string;
}

export interface Edge12 {
  node: Node12;
  __typename: string;
}

export interface Node12 {
  text: Text4;
  __typename: string;
}

export interface Text4 {
  plaidHtml: string;
  __typename: string;
}

export interface QuotesTotal {
  total: number;
  __typename: string;
}

export interface Quotes {
  edges: Edge13[];
  __typename: string;
}
// ...
export interface CrazyCredits {
  edges: Edge14[];
  __typename: string;
}

export interface Edge14 {
  node: Node14;
  __typename: string;
}

export interface Node14 {
  text: Text5;
  __typename: string;
}

export interface Text5 {
  plaidHtml: string;
  __typename: string;
}

export interface AlternateVersions {
  total: number;
  edges: any[];
  __typename: string;
}

export interface Connections {
  edges: Edge15[];
  __typename: string;
}
// ...
export interface Soundtrack {
  edges: Edge16[];
  __typename: string;
}

export interface Edge16 {
  node: Node16;
  __typename: string;
}

export interface Node16 {
  text: string;
  comments: Comment[];
  __typename: string;
}

export interface Comment {
  plaidHtml: string;
  __typename: string;
}

export interface TitleText6 {
  text: string;
  __typename: string;
}

export interface OriginalTitleText6 {
  text: string;
  __typename: string;
}

export interface ReleaseYear5 {
  year: number;
  __typename: string;
}

export interface Reviews2 {
  total: number;
  __typename: string;
}

export interface FeaturedReviews2 {
  edges: Edge17[];
  __typename: string;
}
// ...
export interface Akas {
  edges: Edge21[];
  __typename: string;
}
// ...
export interface Fake {
  "#TITLE": string;
  "#YEAR": number;
  "#IMDB_ID": string;
  "#RANK": number;
  "#ACTORS": string;
  "#AKA": string;
  "#IMDB_URL": string;
  "#IMDB_IV": string;
  "#IMG_POSTER": string;
  photo_width: number;
  photo_height: number;
}

export interface StoryLine {
  id: string;
  summaries: Summaries;
  outlines: Outlines;
  synopses: Synopses;
  storylineKeywords: StorylineKeywords;
  taglines: Taglines;
  genres: Genres2;
  certificate: Certificate3;
  parentsGuide: ParentsGuide;
  __typename: string;
}